examples/acme-industrial/pyproject.toml
[project]
name = "acme-industrial-example"
version = "0.1.0"
description = "Astra — the simulated procurement agent for Acme Industrial."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.0.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.0.0",
    "vouch",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.11",
    "pyright>=1.1.408",
]

[tool.uv]
package = false

[tool.uv.sources]
vouch = { path = "../../sdk-python", editable = true }
